zipkin: remove shared_ptr from c++ interface
c++ shared pointers use atomics for reference counting. passing a
shared_ptr by value to a function incurs an increment on entry and
decrement on return. the ZTracer interface will require many of these
atomic operations over the lifetime of a trace: constructing traces,
constructing child spans, annotations with different endpoints, etc.

because the goal is to use this to measure performance-sensitive code,
including the effect of atomic operations themselves, introduciing more
atomics is really going to get in the way of reliable trace output.

therefore, we remove the factory functions that return shared_ptr, and
all interface functions take raw pointers instead. any lifetime or
reference-counting issues are left for the user to deal with as necessary

Signed-off-by: Casey Bodley <casey@cohortfs.com>